improvement Category Validator
